+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperatorTest.java
@@ -663,16 +663,20 @@ public class KafkaAssemblyOperatorTest {
                 }
 
                 // Check PDBs
-                assertThat(pdbCaptor.getAllValues(), hasSize(3));
+                assertThat(pdbCaptor.getAllValues(), hasSize(4));
                 Set<String> expectedPdbNames = new HashSet<>();
                 expectedPdbNames.add(KafkaResources.kafkaComponentName(CLUSTER_NAME));
                 String entityOperatorName = KafkaResources.entityOperatorDeploymentName(CLUSTER_NAME);
                 String exporterName = KafkaExporterResources.componentName(CLUSTER_NAME);
+                String cruiseControlName = CruiseControlResources.componentName(CLUSTER_NAME);
                 if (kafka.getSpec().getEntityOperator() != null) {
                     expectedPdbNames.add(entityOperatorName);
                 }
                 if (kafka.getSpec().getKafkaExporter() != null) {
                     expectedPdbNames.add(exporterName);
+                }
+                if (kafka.getSpec().getCruiseControl() != null) {
+                    expectedPdbNames.add(cruiseControlName);
                 }
                 assertThat(
                     pdbCaptor.getAllValues().stream()